server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appwrite
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dice-ml==0.11
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
llvmlite==0.42.0
mapie==0.8.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai==1.14.0
packaging==23.2
pandas==2.2.1
Pillow==10.2.0
polars==0.20.15
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
shap==0.45.0
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
xgboost==2.0.3
websockets==12.0
motor==3.3.2
pymongo==4.6.2
